+++ libraries/Microsoft.Bot.Builder.AI.Translation/CustomDictionary.cs
