+++ libraries/Microsoft.Bot.Builder.AI.Translation/DateAndTimeLocaleFormat.cs
