+++ libraries/Microsoft.Bot.Builder.AI.Translation/ILocaleConverter.cs
