+++ libraries/Microsoft.Bot.Builder.AI.Translation/LanguageMap.cs
