+++ libraries/Microsoft.Bot.Builder.AI.Translation/LocaleConverter.cs
