+++ libraries/Microsoft.Bot.Builder.AI.Translation/Microsoft.Bot.Builder.AI.Translation.ruleset
