+++ libraries/Microsoft.Bot.Builder.AI.Translation/Model/Alignment.cs
