+++ libraries/Microsoft.Bot.Builder.AI.Translation/Model/ErrorMessage.cs
