+++ libraries/Microsoft.Bot.Builder.AI.Translation/Model/ErrorModel.cs
