+++ libraries/Microsoft.Bot.Builder.AI.Translation/Model/SentencesLength.cs
