+++ libraries/Microsoft.Bot.Builder.AI.Translation/Model/TranslatedResult.cs
