+++ libraries/Microsoft.Bot.Builder.AI.Translation/Model/TranslationModel.cs
