+++ libraries/Microsoft.Bot.Builder.AI.Translation/Model/TranslatorRequestModel.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Microsoft.Bot.Builder.AI.Translation.Model
+{
+    /// <summary>
+    /// Model for translator request to Translator API v3.
+    /// </summary>
+    internal class TranslatorRequestModel
+    {
+        public string Text { get; set; }
+    }
+}
