+++ libraries/Microsoft.Bot.Builder.AI.Translation/PostProcessor/IPostProcessor.cs
