+++ libraries/Microsoft.Bot.Builder.AI.Translation/PostProcessor/PatternsPostProcessor.cs
