+++ libraries/Microsoft.Bot.Builder.AI.Translation/PostProcessor/PostProcessedDocument.cs
