+++ libraries/Microsoft.Bot.Builder.AI.Translation/PostProcessor/TranslatedDocument.cs
