+++ libraries/Microsoft.Bot.Builder.AI.Translation/TextAndDateTime.cs
